clean the kclhand_contorl pkg
